requirements.txt
Cython==0.29.14
knn-pytorch==1.0
matplotlib==3.1.2
numpy==1.18.0
opencv-python==4.1.2.30
Pillow==6.2.1
pycocotools==2.0.1
PyYAML==5.2
scipy==1.4.1
tqdm==4.41.0